Only signed users can edit and delete books

+++ src/main/java/com/example/demo/Controllers/DeleteBookController.java
@@ -12,11 +12,18 @@ public class DeleteBookController {
     private BookService bookService = new BookService();
 
     @RequestMapping(value = "/books/doDelete/{id}", method = RequestMethod.GET)
-    public String deleteBook(@PathVariable int id, Model model) {
-        bookService.deleteBook(bookService.findBook(id));
-        model.addAttribute("books", bookService.findAllBooks());
-        return "MainView";
+    public String deleteBook(@PathVariable int id,
+                             @ModelAttribute("user") User user,
+                             Model model) {
+        if (user.isRegistered()) {
+            bookService.deleteBook(bookService.findBook(id));
+            model.addAttribute("books", bookService.findAllBooks());
+            return "MainView";
+        } else {
+            return "SignIn";
+        }
     }
+
     @ModelAttribute("user")
     public User setUpUserForm() {
         User user = new User();
